Remove code model recipe module
Those functions are specific to the changeset tooling in Gaphor.
