+++ gaphor/ui/modelmerge/tests/test_modelmerge.py
@@ -1,4 +1,5 @@
-from gaphor.core.modeling import ElementChange, ValueChange, recipes
+from gaphor.core.changeset.compare import set_value_change_property_value
+from gaphor.core.modeling import ElementChange, ValueChange
 from gaphor.ui.modelmerge.editor import ModelMerge
 
 
@@ -13,7 +14,7 @@ def test_build_list_store(event_manager, element_factory, modeling_language):
     vchange.op = "update"
     vchange.element_id = "1234"
     vchange.property_name = "name"
-    recipes.set_value_change_property_value(vchange, "my diagram")
+    set_value_change_property_value(vchange, "my diagram")
 
     model_merge.refresh_model()
